requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
gunicorn==19.7.1
markdown2==2.3.4
Pillow==4.1.1
psycopg2==2.7.3.1
dj-database-url==0.4.2
pytz==2017.2
whitenoise==3.3.1
